refactor : OutputView 를 OutputLadderView 와 OutputResultView로 분리
