+++ src/main/java/view/OutputLadderView.java
@@ -3,18 +3,16 @@ package view;
 import domain.Bridge;
 import domain.BridgeStatus;
 import domain.Ladder;
-import domain.Person;
 
 import java.util.List;
-import java.util.Map;
 
-public class OutputView {
+public class OutputLadderView {
 
     private static final String EMPTY_BRIDGE = "     ";
     private static final String EXIST_BRIDGE = "-----";
-    private static final String RESULT_DELIM = " : ";
-    private static final String TAB = "     ";
     private static final String LADDER_BOUNDARY = "|";
+    private static final String TAB = "     ";
+
 
     public static void printLadder(final Ladder ladder) {
         System.out.println("사다리 결과");
@@ -32,32 +30,6 @@ public class OutputView {
         for (String name : ladder.getParticipantNames()) {
             System.out.print(name + TAB);
         }
-    }
-
-    private static void printResultCandidateOf(final Ladder ladder) {
-        List<String> resultCandidates = ladder.getResultCandidates();
-
-        for (String resultCandidate : resultCandidates) {
-            System.out.print(resultCandidate + TAB);
-        }
-
-        System.out.println();
-    }
-
-    public static void printLadderSpecific(final Map<Person, String> result, final Person person) {
-        System.out.println("실행결과");
-
-        System.out.println(person.getName() + RESULT_DELIM + result.get(person));
-    }
-
-    public static void printLadderAll(final Map<Person, String> result) {
-        for (Person participant : result.keySet()) {
-            System.out.println(participant.getName() + RESULT_DELIM + result.get(participant));
-        }
-    }
-
-    public static void printNotExistedParticipant() {
-        System.out.println("해당 이름을 가진 참여자가 존재하지 않습니다. 다시 입력해 주세요.");
     }
 
     private static void printLinesOf(final Ladder ladder) {
@@ -81,5 +53,15 @@ public class OutputView {
         }
 
         return EXIST_BRIDGE;
+    }
+
+    private static void printResultCandidateOf(final Ladder ladder) {
+        List<String> resultCandidates = ladder.getResultCandidates();
+
+        for (String resultCandidate : resultCandidates) {
+            System.out.print(resultCandidate + TAB);
+        }
+
+        System.out.println();
     }
 }
